requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-core==0.1.16
pydantic==2.5.3
jinja2==3.1.3
